feat(client): code entry inside the settings activation dialog — TelegramLinkFlow extracted and shared with /link
The cross-device path no longer requires leaving settings: the activation
dialog's not-connected state gains 'enter it here →', expanding the same
entry → identity preview (handle + photo) → consequence copy → consented
relink → atomic completion pipeline the /link page runs. One component
(TelegramLinkFlow, page/embedded variants), so there is exactly one
consent surface to keep honest; /link keeps deep-link handling, sign-in
gating, and the page shell. Completing inside the dialog flips the
server-truth link state, which hands over to the existing success splash.
